requirements.txt
anthropic==0.49.0
beautifulsoup4==4.13.4
feedparser==6.0.11
gspread==6.2.0
oauth2client==4.1.3
python-dotenv==1.1.0
Requests==2.32.3
